Made the changes to the shape interface

+++ paint/src/main/java/com/csci4448/paint4448/shapes/Point.java
@@ -1,5 +1,5 @@
 package com.csci4448.paint4448.shapes;
 
 public class Point {
-    public float x, y;
+    public int x, y;
 }
